requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
tensorflow==2.21.0
keras==3.14.0
pillow>=10.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
requests>=2.31.0
playwright>=1.40.0
psycopg2-binary>=2.9.0
